+++ src/syntax/mod.rs
@@ -5,8 +5,8 @@ mod block;
 mod expression;
 mod functions;
 mod params_list;
-mod statement;
 mod utils;
+mod statement;
 
 pub mod ast;