+++ src/syntax/statement.rs
@@ -1,22 +1,19 @@
 use crate::lexic::token::Token;
 
-use super::{ast::statement::Statement, functions::function_call, ParseResult, binding};
+use super::{ast::statement::Statement, binding, functions::function_call, ParseResult};
 
 pub fn try_parse<'a>(tokens: &'a Vec<Token>, pos: usize) -> ParseResult<Statement, ()> {
-    None
-        .or_else(
-            || match binding::try_parse(tokens, pos) {
-                ParseResult::Ok(b, next) => Some(ParseResult::Ok(Statement::Binding(b), next)),
-                ParseResult::Err(err) => Some(ParseResult::Err(err)),
-                _ => None,
-            }
-        )
-        .or_else(|| match function_call::try_parse(tokens, pos) {
-            ParseResult::Ok(f, next) => Some(ParseResult::Ok(Statement::FunctionCall(f), next)),
-            ParseResult::Err(err) => Some(ParseResult::Err(err)),
-            _ => None,
-        })
-        .unwrap_or_else(|| ParseResult::Unmatched)
+    None.or_else(|| match binding::try_parse(tokens, pos) {
+        ParseResult::Ok(b, next) => Some(ParseResult::Ok(Statement::Binding(b), next)),
+        ParseResult::Err(err) => Some(ParseResult::Err(err)),
+        _ => None,
+    })
+    .or_else(|| match function_call::try_parse(tokens, pos) {
+        ParseResult::Ok(f, next) => Some(ParseResult::Ok(Statement::FunctionCall(f), next)),
+        ParseResult::Err(err) => Some(ParseResult::Err(err)),
+        _ => None,
+    })
+    .unwrap_or_else(|| ParseResult::Unmatched)
 }
 
 #[cfg(test)]
@@ -39,7 +36,6 @@ mod tests {
             _ => panic!("Expected a function call"),
         }
     }
-
 
     #[test]
     fn should_parse_binding() {